+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+
+### Fixed
+- Errors in readme.md
+- Weekly interval accepts constructor arguments
+
 ## v1.0.0
 ### Added
 - Base src files, composer, git config, readme and changelog files.